+++ app/src/main/java/com/postpc/tenq/services/RecorderService.java
@@ -11,20 +11,19 @@ import static java.lang.Math.exp;
 
 public class RecorderService {
 
+    public static final int MILLIS = 6000;
     private static double mEMA = 0.0;
     static final private double EMA_FILTER = 0.6;
 
     private MediaRecorder mRecorder;
-    private final Context applicationContext;
     private static Thread runner;
-    private boolean isRecorderOn = true; // TODO: change to false when not debugging
+    private boolean isRecorderOn = false; // need to be true for recording
     private double prevAmplitude;
-    private AudioManager audioManager;
+    private final AudioManager audioManager;
 
     public RecorderService(Context applicationContext) {
 
-        this.applicationContext = applicationContext;
-        this.audioManager = (AudioManager) this.applicationContext.getSystemService(Context.AUDIO_SERVICE);
+        this.audioManager = (AudioManager) applicationContext.getSystemService(Context.AUDIO_SERVICE);
         this.prevAmplitude = getAmplitude();
 
         if (runner == null && isRecorderOn)
@@ -34,35 +33,31 @@ public class RecorderService {
                 {
                     try
                     {
-                        Thread.sleep(6000);
+                        Thread.sleep(MILLIS);
                     }
                     catch (InterruptedException e) {
-                        android.util.Log.e("[Monkey]", "InterruptedException: " + android.util.Log.getStackTraceString(e));
+                        android.util.Log.e("[Recorder]", "InterruptedException: " + android.util.Log.getStackTraceString(e));
                     }
 
                     double currentAmplitude = getAmplitude();
                     int maxVolume = audioManager.getStreamMaxVolume(AudioManager.STREAM_MUSIC);
-                    Log.i("amplitude", String.valueOf(currentAmplitude));
 
                     // TODO: maybe change it to decreasing / increasing by 1 each time the amplitude is lower / higher
                     //  respectively instead of changing the volume using the amplitude value
                     if (abs(currentAmplitude - prevAmplitude) > (double) (32767 / maxVolume)) {
                         int newVolume = (int) (( currentAmplitude / 32767 ) * maxVolume);
 
-                        // add limit so music won't shut down when there are no noises
+                        // add lower bound so music won't shut down when there are no noises
                         if (newVolume > 4) {
                             audioManager.setStreamVolume(AudioManager.STREAM_MUSIC, newVolume, 0);
                         }
                         android.util.Log.e("[volume]", String.valueOf(newVolume));
                     }
                     prevAmplitude = currentAmplitude;
-
-                    Log.i("thread_id", Thread.currentThread().getName());
                     //mHandler.post(runner);
                 }
             });
             runner.start();
-            Log.d("Noise", "start runner()");
         }
     }
 
@@ -85,16 +80,16 @@ public class RecorderService {
             {
                 mRecorder.prepare();
             }catch (java.io.IOException ioe) {
-                android.util.Log.e("[Monkey]", "IOException: " + android.util.Log.getStackTraceString(ioe));
+                android.util.Log.e("[Recorder]", "IOException: " + android.util.Log.getStackTraceString(ioe));
 
             }catch (java.lang.SecurityException e) {
-                android.util.Log.e("[Monkey]", "SecurityException: " + android.util.Log.getStackTraceString(e));
+                android.util.Log.e("[Recorder]", "SecurityException: " + android.util.Log.getStackTraceString(e));
             }
             try
             {
                 mRecorder.start();
             }catch (java.lang.SecurityException e) {
-                android.util.Log.e("[Monkey]", "SecurityException: " + android.util.Log.getStackTraceString(e));
+                android.util.Log.e("[Recorder]", "SecurityException: " + android.util.Log.getStackTraceString(e));
             }
 
         }
