+++ app/src/main/java/com/postpc/tenq/ui/activities/MainActivity.java
@@ -1,21 +1,16 @@
 package com.postpc.tenq.ui.activities;
 
-import android.Manifest;
 import android.app.Activity;
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Bundle;
 import android.util.Log;
 
-import androidx.annotation.NonNull;
-import androidx.core.app.ActivityCompat;
 
 import com.postpc.tenq.R;
 import com.postpc.tenq.core.TenQActivity;
 import com.postpc.tenq.models.User;
 import com.postpc.tenq.network.SpotifyClient;
-import com.postpc.tenq.services.RecorderService;
 import com.spotify.sdk.android.auth.AuthorizationClient;
 import com.spotify.sdk.android.auth.AuthorizationRequest;
 import com.spotify.sdk.android.auth.AuthorizationResponse;
@@ -32,51 +27,14 @@ import static com.postpc.tenq.network.SpotifyClient.CLIENT_ID;
 public class MainActivity extends TenQActivity {
 
     private static final int REQUEST_CODE = 1337;
-    private static RecorderService recorder;
-
-//    @Override
-//    protected void onCreate(Bundle savedInstanceState) {
-//        super.onCreate(savedInstanceState);
-//        setContentView(R.layout.activity_main);
-//
-//        startSpotifyAuthFlow();
-//
-//    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        recorder = new RecorderService(this);
-        if (getAuthService().getCurrentUserId() == null) {
-            startSpotifyAuthFlow();
-        } else {
-            // TODO: maybe we want to move it later
-            // check if the user gave permission to use the microphone
-            if (ActivityCompat.checkSelfPermission(this, Manifest.permission.RECORD_AUDIO) != PackageManager.PERMISSION_GRANTED) {
+        startSpotifyAuthFlow();
 
-                ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.RECORD_AUDIO}, 10);
-
-            } else {
-
-                recorder.startRecorder();
-
-            }
-            startRoomsActivity();
-        }
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (requestCode == 10) {
-            if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                recorder.startRecorder();
-            }else{
-                Log.e("[RECORD_AUDIO]", "Permission denied");
-            }
-        }
     }
 
     private void startRoomsActivity() {
